+++ central/sensor/service/connection/messagestream/stream.go
@@ -1,0 +1,8 @@
+package messagestream
+
+import "github.com/stackrox/rox/generated/internalapi/central"
+
+// CentralMessageStream is a stripped-down version of the SensorService Communicate stream, similar to SensorMessageStream in Sensor.
+type CentralMessageStream interface {
+	Send(msg *central.MsgToSensor) error
+}