+++ sensor/common/networkflow/manager/manager_impl.go
@@ -215,6 +215,8 @@ func NewManager(
 		policyDetector:    policyDetector,
 		enricherTicker:    enricherTicker,
 		finished:          &sync.WaitGroup{},
+		activeConnections: make(map[connection]*networkConnIndicator),
+		activeEndpoints:   make(map[containerEndpoint]*containerEndpointIndicator),
 	}
 
 	if features.SensorCapturesIntermediateEvents.Enabled() {
@@ -238,6 +240,9 @@ type networkFlowManager struct {
 	enrichedConnsLastSentState     map[networkConnIndicator]timestamp.MicroTS
 	enrichedEndpointsLastSentState map[containerEndpointIndicator]timestamp.MicroTS
 	enrichedProcessesLastSentState map[processListeningIndicator]timestamp.MicroTS
+
+	activeConnections map[connection]*networkConnIndicator
+	activeEndpoints   map[containerEndpoint]*containerEndpointIndicator
 
 	done          concurrency.Signal
 	sensorUpdates chan *message.ExpiringMessage
@@ -450,6 +455,12 @@ func (m *networkFlowManager) enrichConnection(conn *connection, status *connStat
 	if !ok {
 		// Expire the connection if the container cannot be found within the clusterEntityResolutionWaitPeriod
 		if timeElapsedSinceFirstSeen > maxContainerResolutionWaitPeriod {
+			if activeConn, found := m.activeConnections[*conn]; found {
+				enrichedConnections[*activeConn] = timestamp.Now()
+				delete(m.activeConnections, *conn)
+				flowMetrics.SetActiveFlowsTotalGauge(len(m.activeConnections))
+				return
+			}
 			status.rotten = true
 			// Only increment metric once the connection is marked rotten
 			flowMetrics.ContainerIDMisses.Inc()
@@ -548,6 +559,15 @@ func (m *networkFlowManager) enrichConnection(conn *connection, status *connStat
 			// hence update the timestamp only if we have a more recent connection than the one we have already enriched.
 			if oldTS, found := enrichedConnections[indicator]; !found || oldTS < status.lastSeen {
 				enrichedConnections[indicator] = status.lastSeen
+				if features.SensorCapturesIntermediateEvents.Enabled() {
+					if status.lastSeen == timestamp.InfiniteFuture {
+						m.activeConnections[*conn] = &indicator
+						flowMetrics.SetActiveFlowsTotalGauge(len(m.activeConnections))
+					} else {
+						delete(m.activeConnections, *conn)
+						flowMetrics.SetActiveFlowsTotalGauge(len(m.activeConnections))
+					}
+				}
 			}
 		}
 	}
@@ -564,6 +584,12 @@ func (m *networkFlowManager) enrichContainerEndpoint(ep *containerEndpoint, stat
 	if !ok {
 		// Expire the connection if the container cannot be found within the clusterEntityResolutionWaitPeriod
 		if timeElapsedSinceFirstSeen > maxContainerResolutionWaitPeriod {
+			if activeEp, found := m.activeEndpoints[*ep]; found {
+				enrichedEndpoints[*activeEp] = timestamp.Now()
+				delete(m.activeEndpoints, *ep)
+				flowMetrics.SetActiveEndpointsTotalGauge(len(m.activeEndpoints))
+				return
+			}
 			status.rotten = true
 			// Only increment metric once the connection is marked rotten
 			flowMetrics.ContainerIDMisses.Inc()
@@ -584,6 +610,15 @@ func (m *networkFlowManager) enrichContainerEndpoint(ep *containerEndpoint, stat
 	// hence update the timestamp only if we have a more recent endpoint than the one we have already enriched.
 	if oldTS, found := enrichedEndpoints[indicator]; !found || oldTS < status.lastSeen {
 		enrichedEndpoints[indicator] = status.lastSeen
+		if features.SensorCapturesIntermediateEvents.Enabled() {
+			if status.lastSeen == timestamp.InfiniteFuture {
+				m.activeEndpoints[*ep] = &indicator
+				flowMetrics.SetActiveEndpointsTotalGauge(len(m.activeEndpoints))
+			} else {
+				delete(m.activeEndpoints, *ep)
+				flowMetrics.SetActiveEndpointsTotalGauge(len(m.activeEndpoints))
+			}
+		}
 	}
 }