+++ ui/apps/platform/src/Containers/Clusters/ClusterEditForm.tsx
@@ -1,4 +1,5 @@
 import React, { ReactElement } from 'react';
+import { Alert, Flex, FlexItem } from '@patternfly/react-core';
 
 import Loader from 'Components/Loader';
 import { labelClassName } from 'constants/form.constants';
@@ -41,7 +42,7 @@ function ClusterEditForm({
     return (
         <div className="bg-base-200 px-4 w-full">
             {/* @TODO, replace open prop with dynamic logic, based on clusterType */}
-            {selectedCluster.id && (
+            {selectedCluster.id ? (
                 <ClusterSummary
                     healthStatus={selectedCluster.healthStatus}
                     status={selectedCluster.status}
@@ -50,6 +51,30 @@ function ClusterEditForm({
                     clusterRetentionInfo={clusterRetentionInfo}
                     isManagerTypeNonConfigurable={isManagerTypeNonConfigurable}
                 />
+            ) : (
+                <Alert variant="warning" isInline title="Legacy installation method" component="p">
+                    <Flex direction={{ default: 'column' }}>
+                        <FlexItem>
+                            <p>
+                                To avoid extra operational complexity, use a{' '}
+                                <strong>cluster init bundle</strong> with either of the following
+                                installation methods:
+                            </p>
+                            <p>
+                                <strong>Operator</strong> for Red Hat OpenShift
+                            </p>
+                            <p>
+                                <strong>Helm chart</strong> for other platforms
+                            </p>
+                        </FlexItem>
+                        <FlexItem>
+                            <p>
+                                Only use the legacy installation method if you have a specific
+                                installation need that requires using this method.
+                            </p>
+                        </FlexItem>
+                    </Flex>
+                </Alert>
             )}
             <form
                 className="grid grid-columns-1 md:grid-columns-2 grid-gap-4 xl:grid-gap-6 mb-4 w-full"